# doc/design/openapi.yaml
openapi: 3.0.3
info:
  title: Shipping on the Air API
  description: Drone-based package delivery system API
  version: 1.0.0

servers:
  - url: http://localhost:8090
    description: Order Service
  - url: http://localhost:8091
    description: Delivery Service
  - url: http://localhost:8092
    description: Drone Service

tags:
  - name: users
    description: User registration and login
  - name: orders
    description: Order management
  - name: deliveries
    description: Delivery management and tracking
  - name: drones
    description: Drone fleet management

paths:

  /api/users/register:
    post:
      tags: [users]
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  username:
                    type: string
        '409':
          description: Username already taken

  /api/users/login:
    post:
      tags: [users]
      summary: Login
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  username:
                    type: string
        '401':
          description: Invalid credentials

  /api/orders:
    post:
      tags: [orders]
      summary: Create a new order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customerId, pickupStreet, pickupLat, pickupLng, deliveryStreet, deliveryLat, deliveryLng, weightKg]
              properties:
                customerId:
                  type: string
                pickupStreet:
                  type: string
                pickupLat:
                  type: number
                pickupLng:
                  type: number
                deliveryStreet:
                  type: string
                deliveryLat:
                  type: number
                deliveryLng:
                  type: number
                weightKg:
                  type: number
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid input

  /api/orders/{id}:
    get:
      tags: [orders]
      summary: Get order details
      operationId: getOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found

  /api/orders/{id}/confirm:
    post:
      tags: [orders]
      summary: Confirm a pending order
      operationId: confirmOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Order cannot be confirmed

  /api/orders/{id}/cancel:
    post:
      tags: [orders]
      summary: Cancel a pending order
      operationId: cancelOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Order cannot be cancelled

  /api/orders/{id}/complete:
    post:
      tags: [orders]
      summary: Mark order as completed (called by Delivery Service)
      operationId: completeOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order completed
        '400':
          description: Order cannot be completed

  /api/deliveries:
    post:
      tags: [deliveries]
      summary: Schedule a new delivery (called by Order Service)
      operationId: scheduleDelivery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderId, pickupLat, pickupLng, deliveryLat, deliveryLng, weightKg]
              properties:
                orderId:
                  type: string
                pickupLat:
                  type: number
                pickupLng:
                  type: number
                deliveryLat:
                  type: number
                deliveryLng:
                  type: number
                weightKg:
                  type: number
      responses:
        '201':
          description: Delivery scheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Delivery'

  /api/deliveries/{id}:
    get:
      tags: [deliveries]
      summary: Track a delivery
      operationId: trackDelivery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Delivery found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Delivery'
        '404':
          description: Delivery not found

  /api/deliveries/{id}/start:
    post:
      tags: [deliveries]
      summary: Start delivery transit
      operationId: startDelivery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Delivery in transit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Delivery'
        '400':
          description: Delivery cannot be started

  /api/deliveries/{id}/complete:
    post:
      tags: [deliveries]
      summary: Complete a delivery
      operationId: completeDelivery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Delivery completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Delivery'
        '400':
          description: Delivery cannot be completed

  /api/deliveries/order/{orderId}:
    get:
      tags: [deliveries]
      summary: Find delivery by order ID
      operationId: findDeliveryByOrderId
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Delivery found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Delivery'
        '404':
          description: Delivery not found

  /api/drones:
    post:
      tags: [drones]
      summary: Register a new drone
      operationId: registerDrone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [maxWeightKg, lat, lng]
              properties:
                maxWeightKg:
                  type: number
                lat:
                  type: number
                lng:
                  type: number
      responses:
        '201':
          description: Drone registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drone'

  /api/drones/available:
    get:
      tags: [drones]
      summary: Find nearest available drone
      operationId: findAvailableDrone
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
        - name: weightKg
          in: query
          required: true
          schema:
            type: number
      responses:
        '200':
          description: Available drone found
          content:
            application/json:
              schema:
                type: object
                properties:
                  droneId:
                    type: string
        '404':
          description: No available drone

  /api/drones/{id}:
    get:
      tags: [drones]
      summary: Get drone details
      operationId: getDrone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Drone found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drone'
        '404':
          description: Drone not found

  /api/drones/{id}/location:
    post:
      tags: [drones]
      summary: Update drone GPS location
      operationId: updateDroneLocation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lat, lng]
              properties:
                lat:
                  type: number
                lng:
                  type: number
      responses:
        '200':
          description: Location updated
        '400':
          description: Drone not found

  /api/drones/{id}/release:
    post:
      tags: [drones]
      summary: Release drone after delivery (called by Delivery Service)
      operationId: releaseDrone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Drone released
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drone'
        '400':
          description: Drone not found

  /api/drones/{id}/assign:
    post:
      tags: [drones]
      summary: Assign drone to a delivery
      operationId: assignDrone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Drone assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drone'
        '400':
          description: Drone not found

components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        pickupStreet:
          type: string
        deliveryStreet:
          type: string
        weightKg:
          type: number
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED]
        createdAt:
          type: string

    Delivery:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        status:
          type: string
          enum: [SCHEDULED, DRONE_ASSIGNED, IN_TRANSIT, DELIVERED, FAILED]
        droneId:
          type: string
        currentLat:
          type: number
        currentLng:
          type: number
        estimatedMinutes:
          type: number
        createdAt:
          type: string

    Drone:
      type: object
      properties:
        id:
          type: string
        maxWeightKg:
          type: number
        lat:
          type: number
        lng:
          type: number
        status:
          type: string
          enum: [AVAILABLE, IN_FLIGHT, CHARGING, MAINTENANCE]